% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDACcurves.R
\name{SDAC_curves}
\alias{SDAC_curves}
\title{SDAC_curves}
\usage{
SDAC_curves(BiddingZone, Date, HourNumber)
}
\arguments{
\item{BiddingZone}{a character-class variable, indicating the bidding zone for which to obtain the curves ("Belgium", "Netherlands", "France", "Germany" or "Austria")}

\item{Date}{a Date-class variable, indicating the date for which to obtain the curves}

\item{HourNumber}{a numeric-class variable, indicating the hour for which to obtain the curves}
}
\value{
a dataframe and ggplot2 figure with the supply and demand curves
}
\description{
SDAC_curves
}
\examples{
curve <- SDAC_curves("Belgium", as.Date("2023-01-01"), 12)
}
